初次合并并修改

+++ src/main/java/InfiniteMusic/entity/User_PlayList.java
@@ -6,11 +6,11 @@ import lombok.Data;
 @Data
 @TableName(value = "user_playlist")
 public class User_PlayList {
-    private int id;//用户id
+    private Long id;//用户id
     private Long playlist_id;
     private boolean create_like;
 
-    public User_PlayList(int userId,long playListId,boolean createOrLike) {
+    public User_PlayList(Long userId,Long playListId,boolean createOrLike) {
         this.id=userId;
         this.playlist_id=playListId;
         this.create_like = createOrLike;
